Add key prefix support to the storage service.
Prior to this change, there was no way for the client to implement key
prefixing on a content-addressed store, since the keys are computed by the CAS
operation on the remote. To support this, add a key prefix field to the message
passed from client to server. On the server side, use this to ensure the
content address gets updated before the blob is written to storage.